+++ src/Core/Domain/Resource/Resource.ts
@@ -1,8 +1,7 @@
 import { Pagination } from "@/Core/Domain/Pagination";
-import { ParsedNumber } from "@/Core/Language";
+import { Maybe, ParsedNumber } from "@/Core/Language";
 import { PageSize } from "../PageSize";
 import { Sort } from "../Sort";
-export * from "./IdParser";
 
 export interface Resource {
   resource_id: string;
@@ -139,4 +138,27 @@ export interface Id {
   agentName: string;
   attribute: string;
   attributeValue: string;
+}
+
+export class IdParser {
+  private static readonly parseIdRegex =
+    /^(?<id>(?<type>(?<ns>[\w-]+(::[\w-]+)*)::(?<class>[\w-]+))\[(?<hostname>[^,]+),(?<attr>[^=]+)=(?<value>[^\]]+)\])(,v=(?<version>[0-9]+))?$/;
+
+  public static parse(idStr: string): Maybe.Maybe<Id> {
+    const groups = idStr.match(IdParser.parseIdRegex)?.groups;
+    if (!groups) {
+      return Maybe.none();
+    }
+    return Maybe.some({
+      entityType: groups.type,
+      agentName: groups.hostname,
+      attribute: groups.attr,
+      attributeValue: groups.value,
+    });
+  }
+
+  public static getAgentName(idStr: string): Maybe.Maybe<Id["agentName"]> {
+    const id = IdParser.parse(idStr);
+    return Maybe.isSome(id) ? Maybe.some(id.value.agentName) : Maybe.none();
+  }
 }
